+++ src/metar.rs
@@ -29,8 +29,8 @@ lazy_static! {
         \s
         (?P<day>\d\d)
         (?P<hour>\d\d)
-        (?P<minute>\d\d)Z?
-        (\s(?P<corrected>COR))?
+        (?P<minute>\d\d)\d?Z?
+        (\s(?P<corrected>COR|CC[A-Z]))?
         (\s(?P<auto>AUTO))?
         (?P<end>\s)
     ").unwrap();
@@ -45,10 +45,10 @@ lazy_static! {
     ").unwrap();
 
     static ref VISIBILITY_RE: Regex = Regex::new(r"(?x)
-        ^(?P<prevailing>[MP]?(\d+\s)?\d/\d|[MP]?\d{1,4}|////|CAVOK)
+        ^(?P<prevailing>[MP]?(\d+\s)?\d/\d{1,2}|[MP]?\d{1,5}|////|[CK]AVOK)
         (NDV)?
         \s?
-        (?P<units>SM)?
+        (?P<units>SM|KM)?
         (\s(?P<minimum>[MP]?\d{1,4}))?
         (?P<directional>(\s[MP]?\d{1,4}[NESW][EW]?)+)?
         (?P<end>\s)
@@ -71,37 +71,48 @@ lazy_static! {
 
     static ref PRESENT_WEATHER_RE: Regex = Regex::new(r"(?x)
         ^(?P<intensity>[-\+])?
-        (?P<vicinity>VC)?
-        (?P<descriptor>MI|BC|PR|DR|BL|SH|TS|FZ)?
-        (?P<phenomena>(DZ|RA|SN|SG|PL|GR|GS|UP|BR|FG|FU|VA|DU|SA|HZ|PO|SQ|FC|SS|DS|IC|PY)+)?
+        (?P<code>(VC|MI|BC|PR|DR|BL|SH|TS|FZ|DZ|RA|SN|SG|PL|GR|GS|UP|BR|FG|FU|VA|DU|SA|HZ|PO|SQ|FC|SS|DS|IC|PY|NSW)+)
         (?P<end>\s)
     ").unwrap();
 
     static ref CLOUD_RE: Regex = Regex::new(r"(?x)
         ^(?P<cover>CLR|SKC|NSC|NCD|FEW|SCT|BKN|OVC|VV|///)
-        (?P<height>\d\d\d|///)?
-        (?P<cloud>AC|ACC|ACSL|AS|CB|CBMAM|CC|CCSL|CI|CS|CU|NS|SC|SCSL|ST|TCU|///)?
+        (?P<height>\d{1,3}|///)?
+        (?P<cloud>AC|ACC|ACSL|AS|CB|CBMAM|CC|CCSL|CI|CS|CU|NS|SC|SCSL|ST|TC?U|///)?
         (?P<end>\s)
     ").unwrap();
 
     static ref TEMPERATURE_RE: Regex = Regex::new(r"(?x)
-        ^(?P<temperature>M?\d\d|//)
+        ^(?P<temperature>M?\d{1,2}|//|XX)
         /
-        (?P<dew_point>M?\d\d|//)?
+        (?P<dew_point>M?\d{1,2}|//|XX)?
         (?P<end>\s)
     ").unwrap();
 
     static ref PRESSURE_RE: Regex = Regex::new(r"(?x)
         ^(?P<units>A|Q)
-        (?P<pressure>\d\d\d\d|////)
+        (?P<pressure>\d{3,4}|////)
         (?P<end>\s)
     ").unwrap();
 
     static ref RECENT_WEATHER_RE: Regex = Regex::new(r"(?x)
         ^RE(?P<intensity>[-\+])?
-        (?P<vicinity>VC)?
-        (?P<descriptor>MI|BC|PR|DR|BL|SH|TS|FZ)?
-        (?P<phenomena>(DZ|RA|SN|SG|PL|GR|GS|UP|BR|FG|FU|VA|DU|SA|HZ|PO|SQ|FC|SS|DS|IC|PY)+)?
+        (?P<code>(VC|MI|BC|PR|DR|BL|SH|TS|FZ|DZ|RA|SN|SG|PL|GR|GS|UP|BR|FG|FU|VA|DU|SA|HZ|PO|SQ|FC|SS|DS|IC|PY|NSW)+)
+        (?P<end>\s)
+    ").unwrap();
+
+    static ref WIND_SHEAR_RE: Regex = Regex::new(r"(?x)
+        ^WS
+        \s
+        (?P<runway>R\d\d[A-Z]?|ALL\sRWY)
+        (?P<end>\s)
+    ").unwrap();
+
+    static ref SEA_RE: Regex = Regex::new(r"(?x)
+        ^W(?P<temperature>M?\d{1,2}|//|XX)
+        /
+        (S(?P<state>\d|/))?
+        (H(?P<height>\d{1,3}|///))?
         (?P<end>\s)
     ").unwrap();
 
@@ -109,11 +120,31 @@ lazy_static! {
         ^(BLACK|BLU\+?|GRN|WHT|RED|AMB|YLO)+
         (?P<end>\s)
     ").unwrap();
+
+    static ref RAINFALL_RE: Regex = Regex::new(r"(?x)
+        ^RF[\d/]{2}[\./][\d/]/[\d/]{3}[\./][\d/]
+        (?P<end>\s)
+    ").unwrap();
+
+    static ref RUNWAY_STATE_RE: Regex = Regex::new(r"(?x)
+        ^R\d\d[A-Z]?/([\d/]{6}|CLRD[\d/]{2})
+        (?P<end>\s)
+    ").unwrap();
+
+    static ref TREND_TIME_RE: Regex = Regex::new(r"(?x)
+        ^(?P<indicator>FM|TL|AT)
+        \s?
+        (?P<hour>\d\d)
+        (?P<minute>\d\d)Z?
+        (?P<end>\s)
+    ").unwrap();
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq)]
+#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
 enum Trend {
+    #[default]
     NoSignificantChange,
     Temporary,
     Becoming,
@@ -133,7 +164,8 @@ impl FromStr for Trend {
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
 enum Section {
     Main,
     Trend(Trend),
@@ -165,7 +197,7 @@ fn handle_section(text: &str) -> Option<(Section, usize)> {
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(tag = "value_type", content = "value", rename_all = "snake_case")]
 enum MetarTime {
     DateTime(UtcDateTime),
@@ -174,20 +206,20 @@ enum MetarTime {
 }
 
 impl MetarTime {
-    fn to_date_time(&self, anchor_time: &NaiveDateTime) -> MetarTime {
+    fn to_date_time(&self, anchor_time: NaiveDateTime) -> MetarTime {
         match self {
             MetarTime::DateTime(utc_dt) => MetarTime::DateTime(*utc_dt),
             MetarTime::DayTime(utc_d_t) => {
                 let first_guess_opt = anchor_time.date().with_day(utc_d_t.0).map(|nd| nd.and_time(utc_d_t.1));
-                let second_guess_opt = (*anchor_time + RelativeDuration::months(-1)).date().with_day(utc_d_t.0).map(|nd| nd.and_time(utc_d_t.1));
-                let third_guess_opt = (*anchor_time + RelativeDuration::months(1)).date().with_day(utc_d_t.0).map(|nd| nd.and_time(utc_d_t.1));
+                let second_guess_opt = (anchor_time + RelativeDuration::months(-1)).date().with_day(utc_d_t.0).map(|nd| nd.and_time(utc_d_t.1));
+                let third_guess_opt = (anchor_time + RelativeDuration::months(1)).date().with_day(utc_d_t.0).map(|nd| nd.and_time(utc_d_t.1));
 
                 let mut final_guess_opt = None;
                 let mut final_delta = i64::MAX;
 
                 for guess_opt in [first_guess_opt, second_guess_opt, third_guess_opt] {
                     if let Some(guess) = guess_opt {
-                        let delta = guess.signed_duration_since(*anchor_time).num_seconds().abs();
+                        let delta = guess.signed_duration_since(anchor_time).num_seconds().abs();
                         if delta < final_delta {
                             final_guess_opt = guess_opt;
                             final_delta = delta;
@@ -197,7 +229,6 @@ impl MetarTime {
 
                 match final_guess_opt {
                     Some(final_guess) => MetarTime::DateTime(UtcDateTime(final_guess)),
-                    // TODO: Make date guessing more robust and correctly handle the error.
                     None => panic!("{}", format!("Date guessing failed, given time {:?} and anchor time {}", self, anchor_time))
                 }
             },
@@ -207,10 +238,10 @@ impl MetarTime {
                 let third_guess = first_guess + Duration::days(1);
 
                 let mut final_guess = first_guess;
-                let mut final_delta = final_guess.signed_duration_since(*anchor_time).num_seconds().abs();
+                let mut final_delta = final_guess.signed_duration_since(anchor_time).num_seconds().abs();
 
                 for guess in [second_guess, third_guess] {
-                    let delta = guess.signed_duration_since(*anchor_time).num_seconds().abs();
+                    let delta = guess.signed_duration_since(anchor_time).num_seconds().abs();
                     if delta < final_delta {
                         final_guess = guess;
                         final_delta = delta;
@@ -224,7 +255,7 @@ impl MetarTime {
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Default, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Default, Serialize, Deserialize)]
 struct Header {
     station_id: Option<String>,
     observation_time: Option<MetarTime>,
@@ -232,7 +263,13 @@ struct Header {
     is_automated: Option<bool>,
 }
 
-fn handle_header(text: &str, anchor_time: Option<&NaiveDateTime>) -> Option<(Header, usize)> {
+impl Header {
+    fn is_empty(&self) -> bool {
+        self.station_id.is_none() && self.observation_time.is_none() && self.is_corrected.is_none() && self.is_automated.is_none()
+    }
+}
+
+fn handle_header(text: &str, anchor_time: Option<NaiveDateTime>) -> Option<(Header, usize)> {
     HEADER_RE.captures(text)
         .map(|capture| {
             let station_id = Some(capture["station_id"].to_string());
@@ -241,7 +278,8 @@ fn handle_header(text: &str, anchor_time: Option<&NaiveDateTime>) -> Option<(Hea
             let hour = capture["hour"].parse().unwrap();
             let minute = capture["minute"].parse().unwrap();
 
-            let mut time = Some(MetarTime::DayTime(UtcDayTime(day, NaiveTime::from_hms_opt(hour, minute, 0).unwrap())));
+            let naive_time = NaiveTime::from_hms_opt(hour, minute, 0);
+            let mut time = naive_time.map(|nt| MetarTime::DayTime(UtcDayTime(day, nt)));
 
             if let Some(at) = anchor_time {
                 time = time.map(|t| t.to_date_time(at));
@@ -260,7 +298,7 @@ fn handle_header(text: &str, anchor_time: Option<&NaiveDateTime>) -> Option<(Hea
 }
 
 #[non_exhaustive]
-#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 enum Unit {
     #[serde(rename = "degT")]
     DegreeTrue,
@@ -268,6 +306,8 @@ enum Unit {
     Knot,
     #[serde(rename = "m/s")]
     MetrePerSecond,
+    #[serde(rename = "km")]
+    KiloMetre,
     #[serde(rename = "m")]
     Metre,
     #[serde(rename = "mi")]
@@ -289,6 +329,7 @@ impl FromStr for Unit {
         match s {
             "KT" => Ok(Unit::Knot),
             "MPS" => Ok(Unit::MetrePerSecond),
+            "KM" => Ok(Unit::KiloMetre),
             "SM" => Ok(Unit::StatuteMile),
             "FT" => Ok(Unit::Foot),
             "Q" => Ok(Unit::HectoPascal),
@@ -320,7 +361,7 @@ fn parse_value(s: &str) -> Result<f32> {
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
 #[serde(tag = "value_type", content = "value", rename_all = "snake_case")]
 enum ValueInRange {
     Above(f32),
@@ -370,7 +411,7 @@ impl Mul<f32> for ValueInRange {
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
 #[serde(tag = "value_type", content = "value", rename_all = "snake_case")]
 enum Value {
     Variable,
@@ -378,8 +419,6 @@ enum Value {
     Below(f32),
     Range(ValueInRange, ValueInRange),
     Exact(f32),
-    Unlimited,
-    Indefinite,
 }
 
 impl FromStr for Value {
@@ -388,17 +427,17 @@ impl FromStr for Value {
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         if s == "VRB" {
             Ok(Value::Variable)
+        } else if s.contains('V') {
+            let mut split = s.split('V');
+            let value1 = ValueInRange::from_str(split.next().unwrap()).unwrap();
+            let value2 = ValueInRange::from_str(split.next().unwrap()).unwrap();
+            Ok(Value::Range(value1, value2))
         } else if let Some(stripped) = s.strip_prefix('P') {
             let value = parse_value(stripped).unwrap();
             Ok(Value::Above(value))
         } else if let Some(stripped) = s.strip_prefix('M') {
             let value = parse_value(stripped).unwrap();
             Ok(Value::Below(value))
-        } else if s.contains('V') {
-            let mut split = s.split('V');
-            let value1 = ValueInRange::from_str(split.next().unwrap()).unwrap();
-            let value2 = ValueInRange::from_str(split.next().unwrap()).unwrap();
-            Ok(Value::Range(value1, value2))
         } else {
             let value = parse_value(s).unwrap();
             Ok(Value::Exact(value))
@@ -416,8 +455,6 @@ impl Div<f32> for Value {
             Value::Below(x) => Value::Below(x / rhs),
             Value::Range(x, y) => Value::Range(x / rhs, y / rhs),
             Value::Exact(x) => Value::Exact(x / rhs),
-            Value::Unlimited => Value::Unlimited,
-            Value::Indefinite => Value::Indefinite,
         }
     }
 }
@@ -432,14 +469,12 @@ impl Mul<f32> for Value {
             Value::Below(x) => Value::Below(x * rhs),
             Value::Range(x, y) => Value::Range(x * rhs, y * rhs),
             Value::Exact(x) => Value::Exact(x * rhs),
-            Value::Unlimited => Value::Unlimited,
-            Value::Indefinite => Value::Indefinite,
         }
     }
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 struct Quantity {
     #[serde(flatten)]
@@ -458,12 +493,18 @@ impl Quantity {
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Default, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Default, Serialize, Deserialize)]
 struct Wind {
     wind_from_direction: Option<Quantity>,
     wind_from_direction_range: Option<Quantity>,
     wind_speed: Option<Quantity>,
     wind_gust: Option<Quantity>,
+}
+
+impl Wind {
+    fn is_empty(&self) -> bool {
+        self.wind_from_direction.is_none() && self.wind_from_direction_range.is_none() && self.wind_speed.is_none() && self.wind_gust.is_none()
+    }
 }
 
 fn handle_wind(text: &str) -> Option<(Wind, usize)> {
@@ -508,7 +549,7 @@ fn handle_wind(text: &str) -> Option<(Wind, usize)> {
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 enum DirectionOctant {
     North,
@@ -540,18 +581,24 @@ impl FromStr for DirectionOctant {
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
 struct DirectionalVisibility {
     visibility: Quantity,
     direction: DirectionOctant,
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Default, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
 struct Visibility {
     prevailing_visibility: Option<Quantity>,
     minimum_visibility: Option<Quantity>,
     directional_visibilites: Vec<DirectionalVisibility>,
+}
+
+impl Visibility {
+    fn is_empty(&self) -> bool {
+        self.prevailing_visibility.is_none() && self.minimum_visibility.is_none() && self.directional_visibilites.is_empty()
+    }
 }
 
 fn handle_visibility(text: &str) -> Option<(Visibility, bool, usize)> {
@@ -561,7 +608,7 @@ fn handle_visibility(text: &str) -> Option<(Visibility, bool, usize)> {
 
             let mut prevailing_visibility_value = match &capture["prevailing"] {
                 "////" => None,
-                "CAVOK" => {
+                "CAVOK" | "KAVOK" => {
                     is_cavok = true;
                     Some(Value::Above(10000.0))
                 },
@@ -601,7 +648,7 @@ fn handle_visibility(text: &str) -> Option<(Visibility, bool, usize)> {
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 enum RunwayVisualRangeTrend {
     Increasing,
@@ -623,7 +670,7 @@ impl FromStr for RunwayVisualRangeTrend {
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 struct RunwayVisualRange {
     runway: String,
     visual_range: Quantity,
@@ -655,7 +702,7 @@ fn handle_runway_visual_range(text: &str) -> Option<(RunwayVisualRange, usize)> 
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 enum WeatherIntensity {
     Light,
@@ -676,7 +723,7 @@ impl FromStr for WeatherIntensity {
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 enum WeatherDescriptor {
     Shallow,
@@ -708,7 +755,7 @@ impl FromStr for WeatherDescriptor {
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 enum WeatherPhenomena {
     Drizzle,
@@ -733,6 +780,7 @@ enum WeatherPhenomena {
     Duststorm,
     IceCrystals,
     Spray,
+    NilSignificantWeather,
 }
 
 impl FromStr for WeatherPhenomena {
@@ -762,17 +810,18 @@ impl FromStr for WeatherPhenomena {
             "DS" => Ok(WeatherPhenomena::Duststorm),
             "IC" => Ok(WeatherPhenomena::IceCrystals),
             "PY" => Ok(WeatherPhenomena::Spray),
+            "NSW" => Ok(WeatherPhenomena::NilSignificantWeather),
             _ => Err(anyhow!("Invalid weather phenomena, given {}", s))
         }
     }
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 struct WeatherCondition {
     intensity: WeatherIntensity,
     is_in_vicinity: bool,
-    descriptor: Option<WeatherDescriptor>,
+    descriptors: Vec<WeatherDescriptor>,
     phenomena: Vec<WeatherPhenomena>,
 }
 
@@ -783,21 +832,33 @@ fn handle_weather(weather_re: &Regex, text: &str) -> Option<(WeatherCondition, u
                 .map(|c| WeatherIntensity::from_str(c.as_str()).unwrap())
                 .unwrap_or(WeatherIntensity::Moderate);
 
-            let is_in_vicinity = capture.name("vicinity").is_some();
-
-            let descriptor = capture.name("descriptor")
-                .map(|c| WeatherDescriptor::from_str(c.as_str()).unwrap());
-
-            let phenomena = capture.name("phenomena")
-                .map(|c| c.as_str().chars().collect::<Vec<_>>()
+            let groups = if &capture["code"] == "NSW" {
+                vec!["NSW".to_string()]
+            } else {
+                capture["code"].chars()
+                    .collect::<Vec<_>>()
                     .chunks(2)
-                    .map(|chunk| WeatherPhenomena::from_str(&chunk.iter().collect::<String>()).unwrap())
-                    .collect())
-                .unwrap_or_default();
+                    .map(String::from_iter)
+                    .collect::<Vec<_>>()
+            };
+
+            let mut is_in_vicinity = false;
+            let mut descriptors = Vec::new();
+            let mut phenomena = Vec::new();
+
+            for group in groups.iter() {
+                if group == "VC" {
+                    is_in_vicinity = true;
+                } else if let Ok(wd) = WeatherDescriptor::from_str(group) {
+                    descriptors.push(wd);
+                } else if let Ok(wp) = WeatherPhenomena::from_str(group) {
+                    phenomena.push(wp);
+                }
+            }
 
             let end = capture.name("end").unwrap().end();
 
-            let weather = WeatherCondition { intensity, is_in_vicinity, descriptor, phenomena };
+            let weather = WeatherCondition { intensity, is_in_vicinity, descriptors, phenomena };
 
             (weather, end)
         })
@@ -812,7 +873,7 @@ fn handle_recent_weather(text: &str) -> Option<(WeatherCondition, usize)> {
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 enum CloudCover {
     Clear,
@@ -847,7 +908,7 @@ impl FromStr for CloudCover {
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 enum CloudType {
     Altocumulus,
@@ -888,14 +949,14 @@ impl FromStr for CloudType {
             "SC" => Ok(CloudType::Stratocumulus),
             "SCSL" => Ok(CloudType::StratocumulusLenticularis),
             "ST" => Ok(CloudType::Stratus),
-            "TCU" => Ok(CloudType::ToweringCumulus),
+            "TCU" | "TU" => Ok(CloudType::ToweringCumulus),
             _ => Err(anyhow!("Invalid cloud type, given {s}"))
         }
     }
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
 struct CloudLayer {
     cover: Option<CloudCover>,
     /// Height above ground level (AGL).
@@ -904,8 +965,8 @@ struct CloudLayer {
 }
 
 impl CloudLayer {
-    fn has_some(&self) -> bool {
-        self.cover.is_some() || self.height.is_some() || self.cloud_type.is_some()
+    fn is_empty(&self) -> bool {
+        self.cover.is_none() && self.height.is_none() && self.cloud_type.is_none()
     }
 }
 
@@ -937,67 +998,29 @@ fn handle_cloud_layer(text: &str) -> Option<(CloudLayer, usize)> {
         })
 }
 
-fn calculate_ceiling(cloud_layers: &[CloudLayer]) -> Option<Quantity> {
-    const UNDEFINED_CEILING: f32 = 999999.9;
-
-    let mut min_ceiling = UNDEFINED_CEILING;
-    let mut is_unlimited_opt = None;
-
-    for cloud_layer in cloud_layers {
-        match cloud_layer.cover {
-            Some(CloudCover::Clear) | Some(CloudCover::SkyClear) | Some(CloudCover::NilSignificantCloud)
-                    | Some(CloudCover::NoCloudDetected) | Some(CloudCover::Few) | Some(CloudCover::Scattered)
-                    | Some(CloudCover::CeilingOk) => {
-                if is_unlimited_opt.is_none() {
-                    is_unlimited_opt = Some(true);
-                }
-            },
-            Some(CloudCover::Broken) | Some(CloudCover::Overcast) | Some(CloudCover::VerticalVisibility) => {
-                if let Some(height_quantity) = &cloud_layer.height {
-                    let height = match height_quantity.value {
-                        Value::Exact(h) => h,
-                        _ => unreachable!(),
-                    };
-
-                    min_ceiling = min_ceiling.min(height);
-                }
-
-                is_unlimited_opt = Some(false);
-            },
-            None => (),
-        }
-    }
-
-    match is_unlimited_opt {
-        Some(is_unlimited) => match is_unlimited {
-            true => Some(Quantity::new(Value::Unlimited, Unit::Foot)),
-            false => if min_ceiling < 999999.0 {
-                Some(Quantity::new(Value::Exact(min_ceiling), Unit::Foot))
-            } else {
-                Some(Quantity::new(Value::Indefinite, Unit::Foot))
-            },
-        }
-        None => None
-    }
-}
-
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Default, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Default, Serialize, Deserialize)]
 struct Temperature {
     temperature: Option<Quantity>,
     dew_point: Option<Quantity>,
+}
+
+impl Temperature {
+    fn is_empty(&self) -> bool {
+        self.temperature.is_none() && self.dew_point.is_none()
+    }
 }
 
 fn handle_temperature(text: &str) -> Option<(Temperature, usize)> {
     TEMPERATURE_RE.captures(text)
         .map(|capture| {
             let temperature_value = match &capture["temperature"] {
-                "//" => None,
+                "//" | "XX" => None,
                 s => Some(Value::from_str(&s.replace('M', "-")).unwrap()),
             };
 
             let dew_point_value = capture.name("dew_point").and_then(|c| match c.as_str() {
-                "//" => None,
+                "//" | "XX" => None,
                 s => Some(Value::from_str(&s.replace('M', "-")).unwrap()),
             });
 
@@ -1013,9 +1036,15 @@ fn handle_temperature(text: &str) -> Option<(Temperature, usize)> {
 }
 
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Default, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Default, Serialize, Deserialize)]
 struct Pressure {
     pressure: Option<Quantity>,
+}
+
+impl Pressure {
+    fn is_empty(&self) -> bool {
+        self.pressure.is_none()
+    }
 }
 
 fn handle_pressure(text: &str) -> Option<(Pressure, usize)> {
@@ -1042,18 +1071,204 @@ fn handle_pressure(text: &str) -> Option<(Pressure, usize)> {
         })
 }
 
-fn handle_color(text: &str) -> Option<usize> {
-    COLOR_RE.captures(text)
+#[non_exhaustive]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+struct WindShear {
+    runway: String,
+}
+
+fn handle_wind_shear(text: &str) -> Option<(WindShear, usize)> {
+    WIND_SHEAR_RE.captures(text)
         .map(|capture| {
+            let runway = match &capture["runway"] {
+                "ALL RWY" => "all".to_string(),
+                s => s[1..].to_string(),
+            };
+
             let end = capture.name("end").unwrap().end();
 
-            end
+            let ws = WindShear { runway };
+
+            (ws, end)
         })
 }
 
-/// A decoded METAR report.
+/// Sea state from WMO Code Table 3700.
 #[non_exhaustive]
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+enum SeaState {
+    Glassy,
+    Rippled,
+    Smooth,
+    Slight,
+    Moderate,
+    Rough,
+    VeryRough,
+    High,
+    VeryHigh,
+    Phenomenal,
+}
+
+impl FromStr for SeaState {
+    type Err = Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "0" => Ok(SeaState::Glassy),
+            "1" => Ok(SeaState::Rippled),
+            "2" => Ok(SeaState::Smooth),
+            "3" => Ok(SeaState::Slight),
+            "4" => Ok(SeaState::Moderate),
+            "5" => Ok(SeaState::Rough),
+            "6" => Ok(SeaState::VeryRough),
+            "7" => Ok(SeaState::High),
+            "8" => Ok(SeaState::VeryHigh),
+            "9" => Ok(SeaState::Phenomenal),
+            _ => Err(anyhow!("Invalid sea state, given {}", s))
+        }
+    }
+}
+
+#[non_exhaustive]
+#[derive(Debug, Clone, Copy, PartialEq, Default, Serialize, Deserialize)]
+struct Sea {
+    sea_temperature: Option<Quantity>,
+    sea_state: Option<SeaState>,
+    wave_height: Option<Quantity>,
+}
+
+impl Sea {
+    fn is_empty(&self) -> bool {
+        self.sea_temperature.is_none() && self.sea_state.is_none() && self.wave_height.is_none()
+    }
+}
+
+fn handle_sea(text: &str) -> Option<(Sea, usize)> {
+    SEA_RE.captures(text)
+        .map(|capture| {
+            let temperature_value = match &capture["temperature"] {
+                "//" | "XX" => None,
+                s => Some(Value::from_str(&s.replace('M', "-")).unwrap()),
+            };
+
+            let sea_state = capture.name("state").and_then(|c| match c.as_str() {
+                "/" => None,
+                s => Some(SeaState::from_str(s).unwrap()),
+            });
+
+            let height_value = capture.name("height").and_then(|c| match c.as_str() {
+                "///" => None,
+                s => Some(Value::from_str(s).unwrap() / 10.0),
+            });
+
+            let sea_temperature = Quantity::new_opt(temperature_value, Unit::DegreeCelsius);
+            let wave_height = Quantity::new_opt(height_value, Unit::Metre);
+
+            let end = capture.name("end").unwrap().end();
+
+            let sea = Sea { sea_temperature, sea_state, wave_height };
+
+            (sea, end)
+        })
+}
+
+fn handle_color(text: &str) -> Option<usize> {
+    COLOR_RE.captures(text)
+        .map(|capture| {
+            capture.name("end").unwrap().end()
+        })
+}
+
+fn handle_rainfall(text: &str) -> Option<usize> {
+    RAINFALL_RE.captures(text)
+        .map(|capture| {
+            capture.name("end").unwrap().end()
+        })
+}
+
+fn handle_runway_state(text: &str) -> Option<usize> {
+    RUNWAY_STATE_RE.captures(text)
+        .map(|capture| {
+            capture.name("end").unwrap().end()
+        })
+}
+
+#[non_exhaustive]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+enum TrendTimeIndicator {
+    From,
+    Until,
+    At,
+}
+
+impl FromStr for TrendTimeIndicator {
+    type Err = Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "FM" => Ok(TrendTimeIndicator::From),
+            "TL" => Ok(TrendTimeIndicator::Until),
+            "AT" => Ok(TrendTimeIndicator::At),
+            _ => Err(anyhow!("Invalid trend time indicator, given {}", s))
+        }
+    }
+}
+
+#[non_exhaustive]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+struct TrendTime {
+    indicator: TrendTimeIndicator,
+    time: Option<MetarTime>,
+}
+
+fn handle_trend_time(text: &str, anchor_time: Option<NaiveDateTime>) -> Option<(TrendTime, usize)> {
+    TREND_TIME_RE.captures(text)
+        .map(|capture| {
+            let indicator = TrendTimeIndicator::from_str(&capture["indicator"]).unwrap();
+            let mut hour = capture["hour"].parse().unwrap();
+            let minute = capture["minute"].parse().unwrap();
+
+            if hour == 24 {
+                hour = 0;
+            }
+
+            let naive_time = NaiveTime::from_hms_opt(hour, minute, 0);
+            let mut time = naive_time.map(|nt| MetarTime::Time(UtcTime(nt)));
+
+            if let Some(at) = anchor_time {
+                time = time.map(|t| t.to_date_time(at));
+            }
+
+            let end = capture.name("end").unwrap().end();
+
+            let trend_time = TrendTime { indicator, time };
+
+            (trend_time, end)
+        })
+}
+
+/// Significant changes in the meteorological conditions in the TREND forecast.
+/// Only elements for which a significant change is expected are [Option::Some].
+#[non_exhaustive]
+#[derive(Debug, Clone, Default, PartialEq, Serialize, Deserialize)]
+struct TrendChange {
+    indicator: Trend,
+    from_time: Option<MetarTime>,
+    to_time: Option<MetarTime>,
+    at_time: Option<MetarTime>,
+    #[serde(flatten)]
+    wind: Wind,
+    #[serde(flatten)]
+    visibility: Visibility,
+    weather: Vec<WeatherCondition>,
+    clouds: Vec<CloudLayer>,
+}
+
+/// Decoded METAR report.
+#[non_exhaustive]
+#[derive(Debug, Clone, Default, PartialEq, Serialize, Deserialize)]
 pub struct Metar {
     #[serde(flatten)]
     header: Header,
@@ -1064,12 +1279,15 @@ pub struct Metar {
     runway_visual_ranges: Vec<RunwayVisualRange>,
     present_weather: Vec<WeatherCondition>,
     clouds: Vec<CloudLayer>,
-    ceiling: Option<Quantity>,
     #[serde(flatten)]
     temperature: Temperature,
     #[serde(flatten)]
     pressure: Pressure,
     recent_weather: Vec<WeatherCondition>,
+    wind_shears: Vec<WindShear>,
+    #[serde(flatten)]
+    sea: Sea,
+    trend_changes: Vec<TrendChange>,
     pub report: String,
 }
 
@@ -1078,21 +1296,18 @@ pub struct Metar {
 /// The optional `anchor_time` specifies a day close the the one when the report was collected.
 /// If given, the decoded METAR day and time will be matched against it to create [UtcDateTime]
 /// struct which fully describes date and time.
-pub fn decode_metar(report: &str, anchor_time: Option<&NaiveDateTime>) -> Result<Metar> {
+pub fn decode_metar(report: &str, anchor_time: Option<NaiveDateTime>) -> Result<Metar> {
     let mut sanitized = report.to_uppercase().trim().replace('\x00', "");
     sanitized = WHITESPACE_REPLACE_RE.replace_all(&sanitized, *WHITESPACE_REPLACE_OUT).to_string();
     let report = END_REPLACE_RE.replace_all(&sanitized, *END_REPLACE_OUT).to_string();
 
     let mut section = Section::Main;
-    let mut header = None;
-    let mut wind = None;
-    let mut visibility = None;
-    let mut runway_visual_ranges = Vec::new();
-    let mut present_weather_conditions = Vec::new();
-    let mut clouds = Vec::new();
-    let mut temperature = None;
-    let mut pressure = None;
-    let mut recent_weather_conditions = Vec::new();
+
+    let mut metar = Metar::default();
+    metar.report = report.trim().to_string();
+
+    let mut processing_trend_change = false;
+    let mut trend_change = TrendChange::default();
 
     let mut unparsed_groups = Vec::new();
 
@@ -1104,34 +1319,46 @@ pub fn decode_metar(report: &str, anchor_time: Option<&NaiveDateTime>) -> Result
         if let Some((sec, relative_end)) = handle_section(sub_report) {
             section = sec;
             idx += relative_end;
+
+            if processing_trend_change {
+                metar.trend_changes.push(trend_change.clone());
+                processing_trend_change = false;
+                trend_change = TrendChange::default();
+            }
+
+            if let Section::Trend(trend) = section {
+                processing_trend_change = true;
+                trend_change.indicator = trend;
+            }
+
             continue;
         }
 
         match section {
             Section::Main => {
-                if header.is_none() {
-                    if let Some((h, relative_end)) = handle_header(sub_report, anchor_time) {
-                        header = Some(h);
+                if metar.header.is_empty() {
+                    if let Some((header, relative_end)) = handle_header(sub_report, anchor_time) {
+                        metar.header = header;
                         idx += relative_end;
                         continue;
                     }
                 }
 
-                if wind.is_none() {
-                    if let Some((w, relative_end)) = handle_wind(sub_report) {
-                        wind = Some(w);
+                if metar.wind.is_empty() {
+                    if let Some((wind, relative_end)) = handle_wind(sub_report) {
+                        metar.wind = wind;
                         idx += relative_end;
                         continue;
                     }
                 }
 
-                if visibility.is_none() {
-                    if let Some((vis, is_cavok, relative_end)) = handle_visibility(sub_report) {
-                        visibility = Some(vis);
+                if metar.visibility.is_empty() {
+                    if let Some((visibility, is_cavok, relative_end)) = handle_visibility(sub_report) {
+                        metar.visibility = visibility;
 
                         if is_cavok {
-                            let cl = CloudLayer { cover: Some(CloudCover::CeilingOk) , height: None, cloud_type: None };
-                            clouds.push(cl);
+                            let cloud_layer = CloudLayer { cover: Some(CloudCover::CeilingOk) , height: None, cloud_type: None };
+                            metar.clouds.push(cloud_layer);
                         }
 
                         idx += relative_end;
@@ -1139,88 +1366,165 @@ pub fn decode_metar(report: &str, anchor_time: Option<&NaiveDateTime>) -> Result
                     }
                 }
 
-                if let Some((pw, relative_end)) = handle_present_weather(sub_report) {
-                    present_weather_conditions.push(pw);
+                if let Some((weather_condition, relative_end)) = handle_present_weather(sub_report) {
+                    metar.present_weather.push(weather_condition);
                     idx += relative_end;
                     continue;
                 }
 
-                if let Some((rvr, relative_end)) = handle_runway_visual_range(sub_report) {
-                    runway_visual_ranges.push(rvr);
+                if let Some((runway_visual_range, relative_end)) = handle_runway_visual_range(sub_report) {
+                    metar.runway_visual_ranges.push(runway_visual_range);
                     idx += relative_end;
                     continue;
                 }
 
-                if let Some((cl, relative_end)) = handle_cloud_layer(sub_report) {
-                    if cl.has_some() {
-                        clouds.push(cl);
+                if let Some((cloud_layer, relative_end)) = handle_cloud_layer(sub_report) {
+                    if !cloud_layer.is_empty() {
+                        metar.clouds.push(cloud_layer);
                     }
+
                     idx += relative_end;
                     continue;
                 }
 
-                if temperature.is_none() {
-                    if let Some((temp, relative_end)) = handle_temperature(sub_report) {
-                        temperature = Some(temp);
+                if metar.temperature.is_empty() {
+                    if let Some((temperature, relative_end)) = handle_temperature(sub_report) {
+                        if !temperature.is_empty() {
+                            metar.temperature = temperature;
+                        }
+
                         idx += relative_end;
                         continue;
                     }
                 }
 
-                if let Some((p, relative_end)) = handle_pressure(sub_report) {
-                    if pressure.is_none() {
-                        pressure = Some(p);
+                if metar.pressure.is_empty() {
+                    if let Some((pressure, relative_end)) = handle_pressure(sub_report) {
+                        metar.pressure = pressure;
+                        idx += relative_end;
+                        continue;
                     }
+                }
+
+                if let Some((weather_condition, relative_end)) = handle_recent_weather(sub_report) {
+                    metar.recent_weather.push(weather_condition);
                     idx += relative_end;
                     continue;
                 }
 
-                if let Some((rw, relative_end)) = handle_recent_weather(sub_report) {
-                    recent_weather_conditions.push(rw);
+                if let Some((wind_shear, relative_end)) = handle_wind_shear(sub_report) {
+                    metar.wind_shears.push(wind_shear);
                     idx += relative_end;
                     continue;
                 }
 
+                if metar.sea.is_empty() {
+                    if let Some((sea, relative_end)) = handle_sea(sub_report) {
+                        if !sea.is_empty() {
+                            metar.sea = sea;
+                        }
+
+                        idx += relative_end;
+                        continue;
+                    }
+                }
+
+                // Colour state, won't store. For more info check:
+                // <https://en.wikipedia.org/wiki/Colour_state>
                 if let Some(relative_end) = handle_color(sub_report) {
                     idx += relative_end;
                     continue;
                 }
+
+                // Rainfall in last 10min / since 0900 local time, won't store. For more info check:
+                // <http://www.bom.gov.au/aviation/Aerodrome/metar-speci.pdf>
+                if let Some(relative_end) = handle_rainfall(sub_report) {
+                    idx += relative_end;
+                    continue;
+                }
+
+                // Runway state (should be part of SNOWTAM), won't store. For more info check:
+                // <https://www.icao.int/WACAF/Documents/Meetings/2021/GRF/2.%20Provisions%20on%20GRF.pdf>
+                if let Some(relative_end) = handle_runway_state(sub_report) {
+                    idx += relative_end;
+                    continue;
+                }
             },
-            Section::Trend(_) => (), // TODO: https://github.com/meandair/rweather-decoder/issues/14
+            Section::Trend(_) => {
+                if let Some((trend_time, relative_end)) = handle_trend_time(sub_report, anchor_time) {
+                    match trend_time.indicator {
+                        TrendTimeIndicator::From => {
+                            trend_change.from_time = trend_time.time;
+                        },
+                        TrendTimeIndicator::Until => {
+                            trend_change.to_time = trend_time.time;
+                        },
+                        TrendTimeIndicator::At => {
+                            trend_change.at_time = trend_time.time;
+                        },
+                    }
+
+                    idx += relative_end;
+                    continue;
+                }
+
+                if trend_change.wind.is_empty() {
+                    if let Some((wind, relative_end)) = handle_wind(sub_report) {
+                        trend_change.wind = wind;
+                        idx += relative_end;
+                        continue;
+                    }
+                }
+
+                if trend_change.visibility.is_empty() {
+                    if let Some((visibility, is_cavok, relative_end)) = handle_visibility(sub_report) {
+                        trend_change.visibility = visibility;
+
+                        if is_cavok {
+                            let cloud_layer = CloudLayer { cover: Some(CloudCover::CeilingOk) , height: None, cloud_type: None };
+                            trend_change.clouds.push(cloud_layer);
+                        }
+
+                        idx += relative_end;
+                        continue;
+                    }
+                }
+
+                if let Some((weather_condition, relative_end)) = handle_present_weather(sub_report) {
+                    trend_change.weather.push(weather_condition);
+                    idx += relative_end;
+                    continue;
+                }
+
+                if let Some((cloud_layer, relative_end)) = handle_cloud_layer(sub_report) {
+                    if !cloud_layer.is_empty() {
+                        trend_change.clouds.push(cloud_layer);
+                    }
+
+                    idx += relative_end;
+                    continue;
+                }
+            },
             Section::Remark => (), // TODO: https://github.com/meandair/rweather-decoder/issues/15
         }
 
         let relative_end = sub_report.find(' ').unwrap();
 
-        if section == Section::Main { // TODO: Push from all sections that are being decoded.
-            let unparsed = &report[idx..idx + relative_end];
-            if unparsed.chars().any(|c| c != '/') {
-                unparsed_groups.push(unparsed);
-            }
+        let unparsed = &report[idx..idx + relative_end];
+        if unparsed.chars().any(|c| c != '/') {
+            unparsed_groups.push(unparsed);
         }
 
         idx += relative_end + 1;
     }
 
-    let report = report.trim().to_string();
+    if processing_trend_change {
+        metar.trend_changes.push(trend_change);
+    }
 
     if !unparsed_groups.is_empty() {
         debug!("Unparsed data: {}, report: {}", unparsed_groups.join(" "), report);
     }
 
-    let ceiling = calculate_ceiling(&clouds);
-
-    Ok(Metar {
-        header: header.unwrap_or_default(),
-        wind: wind.unwrap_or_default(),
-        visibility: visibility.unwrap_or_default(),
-        present_weather: present_weather_conditions,
-        runway_visual_ranges,
-        clouds,
-        ceiling,
-        temperature: temperature.unwrap_or_default(),
-        pressure: pressure.unwrap_or_default(),
-        recent_weather: recent_weather_conditions,
-        report,
-    })
+    Ok(metar)
 }